+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - READ.ME to add instructions for [Balatest](https://github.com/BakersDozenBagels/Balatest/) (#13)
 ### Fixed
 - Fixed Blueprint/Brainstorm being parsed incorrectly when the mod Blueprint is installed. (#5)
+- Discard amount not being parsed with Banner Joker (#14)
 - GameState not removing a joker when Madness is activated (#15)
 
 ### Removed